change onupdate and onstop into double

+++ src/main/java/com/next/pong/framework/activity/Activity.java
@@ -31,7 +31,7 @@ public abstract class Activity<T extends Layout> {
         return payload;
     }
 
-    public void onUpdate(float deltaTime) {
+    public void onUpdate(double deltaTime) {
         layout.onUpdate(deltaTime);
     }
